refactor: :truck: rename file-utils to import-utils
file-utils was only used for file import
