+++ src/screens/file-selection-screen/utils/__tests__/import-utils.test.ts
@@ -1,4 +1,4 @@
-import { handleFileImport } from '../file-utils';
+import { handleFileImport } from '../import-utils';
 
 afterEach(() => {
   jest.clearAllMocks();
